WIP: Add a "plugin finder" layer of indirection
