+++ samples/android/app/build.gradle
@@ -54,5 +54,5 @@ ext.gleanNamespace = "mozilla.telemetry.glean"
 // Include the glean-gradle-plugin. This is slightly different than what is
 // recommended for external users since we are loading it from the same root Gradle
 // build.
-apply from: '../../../gradle-plugin/src/main/groovy/mozilla/telemetry/glean-gradle-plugin/GleanGradlePlugin.groovy'
+apply from: '../../../gradle-plugin/GleanGradlePluginCore.groovy'
 ext.glean_plugin.apply(project)
